Pull diagnostics out to telemetry.py, add expo backoff, add aiofiles to NOTICE
